Separate out interceptorbatchmethodimpl to a different file
